separate app encoding and params
