+++ app/consumer/ante/msg_filter_ante_test.go
@@ -10,7 +10,7 @@ import (
 	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
 
 	"github.com/cosmos/interchain-security/v4/app/consumer/ante"
-	appparams "github.com/cosmos/interchain-security/v4/app/consumerparams"
+	appencoding "github.com/cosmos/interchain-security/v4/app/encoding"
 )
 
 type consumerKeeper struct {
@@ -28,7 +28,7 @@ func noOpAnteDecorator() sdk.AnteHandler {
 }
 
 func TestMsgFilterDecorator(t *testing.T) {
-	txCfg := appparams.MakeTestEncodingConfig().TxConfig
+	txCfg := appencoding.MakeTestEncodingConfig().TxConfig
 
 	testCases := []struct {
 		name           string
