--- app/consumerparams/params.go
@@ -1,7 +1,0 @@
-package consumerparams
-
-// Simulation parameter constants
-const (
-	StakePerAccount           = "stake_per_account"
-	InitiallyBondedValidators = "initially_bonded_validators"
-)
