--- app/consumerparams/weights.go
@@ -1,42 +1,0 @@
-package consumerparams
-
-// Default simulation operation weights for messages and gov proposals
-const (
-	DefaultWeightMsgSend                        int = 100
-	DefaultWeightMsgMultiSend                   int = 10
-	DefaultWeightMsgSetWithdrawAddress          int = 50
-	DefaultWeightMsgWithdrawDelegationReward    int = 50
-	DefaultWeightMsgWithdrawValidatorCommission int = 50
-	DefaultWeightMsgFundCommunityPool           int = 50
-	DefaultWeightMsgDeposit                     int = 100
-	DefaultWeightMsgVote                        int = 67
-	DefaultWeightMsgUnjail                      int = 100
-	DefaultWeightMsgCreateValidator             int = 100
-	DefaultWeightMsgEditValidator               int = 5
-	DefaultWeightMsgDelegate                    int = 100
-	DefaultWeightMsgUndelegate                  int = 100
-	DefaultWeightMsgBeginRedelegate             int = 100
-
-	DefaultWeightCommunitySpendProposal int = 5
-	DefaultWeightTextProposal           int = 5
-	DefaultWeightParamChangeProposal    int = 5
-
-	DefaultWeightMsgStoreCode           int = 50
-	DefaultWeightMsgInstantiateContract int = 100
-	DefaultWeightMsgExecuteContract     int = 100
-	DefaultWeightMsgUpdateAdmin         int = 25
-	DefaultWeightMsgClearAdmin          int = 10
-	DefaultWeightMsgMigrateContract     int = 50
-
-	DefaultWeightStoreCodeProposal                   int = 5
-	DefaultWeightInstantiateContractProposal         int = 5
-	DefaultWeightUpdateAdminProposal                 int = 5
-	DefaultWeightExecuteContractProposal             int = 5
-	DefaultWeightClearAdminProposal                  int = 5
-	DefaultWeightMigrateContractProposal             int = 5
-	DefaultWeightSudoContractProposal                int = 5
-	DefaultWeightPinCodesProposal                    int = 5
-	DefaultWeightUnpinCodesProposal                  int = 5
-	DefaultWeightUpdateInstantiateConfigProposal     int = 5
-	DefaultWeightStoreAndInstantiateContractProposal int = 5
-)
