--- app/providerparams/params.go
@@ -1,7 +1,0 @@
-package providerparams
-
-// Simulation parameter constants
-const (
-	StakePerAccount           = "stake_per_account"
-	InitiallyBondedValidators = "initially_bonded_validators"
-)
