+++ cmd/interchain-security-cdd/main.go
@@ -7,12 +7,13 @@ import (
 	svrcmd "github.com/cosmos/cosmos-sdk/server/cmd"
 
 	app "github.com/cosmos/interchain-security/v4/app/consumer-democracy"
+	appparams "github.com/cosmos/interchain-security/v4/app/params"
 	"github.com/cosmos/interchain-security/v4/cmd/interchain-security-cdd/cmd"
 )
 
 func main() {
+	appparams.SetAddressPrefixes("consumer")
 	rootCmd := cmd.NewRootCmd()
-
 	if err := svrcmd.Execute(rootCmd, "", app.DefaultNodeHome); err != nil {
 		switch e := err.(type) {
 		case server.ErrorCode:
